refactor(interpreter): modify the interpreter to eval imported packages each time
The previous method for the interpreter was to eval each package and
then keep the package for future use. But each usage involved creating a
clone of the package and it would involve fixing the objects to refer to
the new package objects in order to keep things like options using the
original reference value.

This modifies the workflow so imports are evaluated each time instead of
attempting to share the values. This greatly simplifies the interpreter
code because we no longer have to clone values before execution and
worry about whether we are referring to the correct object.